Set PREFECT_API_URL correctly during tests
