+++ backend/tests/helpers/utils.py
@@ -1,11 +1,12 @@
 from contextlib import contextmanager
 from typing import Generator
 
+import pytest
 from testcontainers.core.container import DockerContainer
 from testcontainers.core.waiting_utils import wait_for_logs
 
 from infrahub.services import InfrahubServices, services
-from tests.helpers.constants import PORT_BOLT_NEO4J, PORT_HTTP_NEO4J
+from tests.helpers.constants import INFRAHUB_USE_TEST_CONTAINERS, PORT_BOLT_NEO4J, PORT_HTTP_NEO4J, PORT_PREFECT
 
 
 def get_exposed_port(container: DockerContainer, port: int) -> int:
@@ -33,6 +34,28 @@ def start_neo4j_container(neo4j_image: str) -> DockerContainer:
     container.start()
     wait_for_logs(container, "Started.")  # wait_container_is_ready does not seem to be enough
     return container
+
+
+def start_prefect_server_container(
+    request: pytest.FixtureRequest,
+) -> dict[int, int] | None:
+    if not INFRAHUB_USE_TEST_CONTAINERS:
+        return None
+
+    container = (
+        DockerContainer(image="prefecthq/prefect:3.0.11-python3.12")
+        .with_command("prefect server start --host 0.0.0.0 --ui")
+        .with_exposed_ports(PORT_PREFECT)
+    )
+
+    def cleanup() -> None:
+        container.stop()
+
+    container.start()
+    wait_for_logs(container, "Configure Prefect to communicate with the server")
+    request.addfinalizer(cleanup)
+
+    return {PORT_PREFECT: get_exposed_port(container, PORT_PREFECT)}
 
 
 @contextmanager
